split up files to reduce potential conflicts since everyone gets a separate file to work on and two or more people dont work on the same files at the same time
