+++ templates/cases_and_vaccines.html
@@ -74,5 +74,7 @@
   </div> -->
   </div>
 </div>
+
+<script src="static/js/scatter_by_nicolee.js"></script>
 <!-- Rest of your content here -->
 {% endblock %}
