+++ templates/health.html
@@ -7,4 +7,7 @@
 
 health goes here
 <!-- Rest of your content here -->
+
+<script src="static/js/health.js"></script>
+
 {% endblock %}
