+++ templates/migration.html
@@ -7,4 +7,6 @@
   migration goes 
 
 <!-- Rest of your content here -->
+
+<script src="static/js/migration.js"></script>
 {% endblock %}
